checkin opengraph refactoring
