+++ src/pages/404.js
@@ -1,23 +1,31 @@
 import React from "react"
-import SEO from "../components/seo"
+import { Helmet } from "react-helmet"
 import Layout from "../components/layout"
 import "../sass/404.scss"
 
-const NotFoundPage = () => (
-  <Layout>
-    <div>
-      <SEO title="404: Not found" />
+const NotFoundPage = () => {
+  const notFoundTitle = "404: Not found"
+  const notFoundDescription =
+    "You just hit a route that doesnt exist... oh, the sadness."
+
+  return (
+    <Layout>
+      <Helmet>
+        <html lang="en" />
+        <title>{notFoundTitle}</title>
+        <meta name="description" content={notFoundDescription} />
+      </Helmet>
       <div className="nonexistent-header"></div>
       <div className="not-found">
         <div className="nonexistent-container">
           <div className="nonexistent-content">
-            <h1>404: NOT FOUND</h1>
-            <h4>You just hit a route that doesnt exist... oh, the sadness.</h4>
+            <h1>{notFoundTitle}</h1>
+            <h4>{notFoundDescription}</h4>
           </div>
         </div>
       </div>
-    </div>
-  </Layout>
-)
+    </Layout>
+  )
+}
 
 export default NotFoundPage
